+++ CHANGES.md
@@ -1,10 +1,12 @@
 
+- Build configuration key parsing. Fail hard on any error instead
+  of warning (#56)
 - Add a `--debug` configuration key. Defaults to `true` or the value
   of the `TOPKG_CONF_DEBUG` environment variable. The default build
   system invocation is changed to enable save of debugging information
   in build artefacts if they key is `true`. The key is generally not
   meant to be specified so that the policy can changed in bulk over
-  topkg packages.
+  topkg packages (#54).
 - `Pkg.files_to_watermark`, make sure only files are returned (#58).
 - Improve error message of some `Topkg.OS` functions (#57).
 - Remove deprecated `--installer` configuration key.